// src/generated/serverApi/v3/models/course-metadata-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CourseMetadataResponse
 */
export interface CourseMetadataResponse {
    /**
     * The id of the Grid element
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    id: string;
    /**
     * Title of the Grid element
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    title: string;
    /**
     * Short title of the Grid element
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    shortTitle: string;
    /**
     * Color of the Grid element
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    displayColor: string;
    /**
     * Indicates if the course is locked and cannot be accessed.
     * @type {boolean}
     * @memberof CourseMetadataResponse
     */
    isLocked: boolean;
    /**
     * Start date of the course
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    startDate?: string;
    /**
     * End date of the course. After this the course counts as archived
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    untilDate?: string;
    /**
     * Start of the copying process if it is still ongoing - otherwise property is not set.
     * @type {string}
     * @memberof CourseMetadataResponse
     */
    copyingSince?: string;
}